pyproject.toml
[project]
name = "emoconnect"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "django-allauth[socialaccount]>=65.11.1",
    "dj-rest-auth>=7.0.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-yasg>=1.21.10",
    "whitenoise>=6.10.0",
    "python-decouple>=3.8",
    "gunicorn>=23.0.0",
    "black>=25.9.0",
    "isort>=6.0.1",
]
